Add a privacy policy
Given that we publish things like raw survey results, we should have an explicit privacy policy to ensure that people know that if they proactively provide PII, then that information could be made public in our GitHub repo and other materials.

Also, in future surveys, we should be clear about this fact.